+++ Micro service/FoundationService/Awine.Teach.FoundationService/Application/ViewModels/TenantSettings/TenantSettingsAddViewModel.cs
@@ -38,6 +38,6 @@ namespace Awine.Teach.FoundationService.Application.ViewModels
         /// <summary>
         /// 最大存储空间
         /// </summary>
-        public int MaxStorageSpace { get; set; }
+        public long MaxStorageSpace { get; set; }
     }
 }
